Refactor HomePage layout for better responsiveness; update ProductView component with loading and error states, and change price type to float in products migration

+++ server/database/migrations/2024_11_18_171158_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable();
-            $table->integer('price')->nullable();
+            $table->float('price')->nullable();
             $table->string('description')->nullable();
             $table->boolean('status')->default(1);
             $table->string('image')->nullable();
